+++ lib/rrj/tools/gem/config/rabbit.rb
@@ -23,11 +23,19 @@ module RubyRabbitmqJanus
       # @return [Hash] Format hash for bunny settings
       def server_settings
         Hash[%w[host port pass user vhost log_level].map do |value|
-          [
-            value.to_sym,
-            @options['rabbit'][value.eql?('log_level') ? 'level' : value]
-          ]
+          key = value.to_sym
+          j_value = @options['rabbit'][rabbitmq_conf(value)]
+
+          raise Errors::Tools::Config::Rabbitmq value if j_value.blank?
+
+          [key, j_value]
         end]
+      end
+
+      private
+
+      def rabbitmq_conf(value)
+        value.eql?('log_level') ? 'level' : value
       end
     end
   end